remove old Jython hacks (#3312)
* remove old Jython hacks

* I fail to see how this is clearer than what we had
