+++ Bio/Phylo/PAML/_parse_codeml.py
@@ -309,8 +309,6 @@ def parse_model(lines, results):
             branch = branch_res.group(1)
             if parameters.get("branches") is None:
                 parameters["branches"] = {}
-            # Hack for Jython http://bugs.jython.org/issue1762 float("-nan")
-            line = line.replace(" -nan", " nan")
             params = line.strip().split()[1:]
             parameters["branches"][branch] = {
                 "t": float(params[0].strip()),
