+++ Bio/PopGen/GenePop/Controller.py
@@ -160,11 +160,7 @@ class _FileIterator:
 
     def __del__(self):
         self.stream.close()
-        try:
-            os.remove(self.fname)
-        except OSError:
-            # Jython seems to call the iterator twice
-            pass
+        os.remove(self.fname)
 
 
 class _GenePopCommandline(AbstractCommandline):
