+++ Tests/test_NCBITextParser.py
@@ -730,15 +730,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.database_letters, 29652561)
         self.assertEqual(len(record.rounds), 3)
         self.assertEqual(len(record.rounds[0].new_seqs), 14)
-        # Rest of test broken up to avoid Jython JVM limitations
-        self._check_text_2010L_phiblast_002_round0(record)
-        self._check_text_2010L_phiblast_002_round1(record)
-        self._check_text_2010L_phiblast_002_round2(record)
-        self._check_text_2010L_phiblast_002_hsps(record)
-        self._check_text_2010L_phiblast_002_hsps_details(record)
-        self._check_text_2010L_phiblast_002_footer(record)
-
-    def _check_text_2010L_phiblast_002_round0(self, record):
         self.assertEqual(record.rounds[0].new_seqs[0].title, "gi|126343|sp|P17216|LIVK_SALTY LEUCINE-SPECIFIC BINDING PROTEIN...")
         self.assertEqual(record.rounds[0].new_seqs[0].score, 743)
         self.assertAlmostEqual(record.rounds[0].new_seqs[0].e, 0.0)
@@ -810,8 +801,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[0].alignments[12].length, 542)
         self.assertEqual(record.rounds[0].alignments[13].title, ">gi|1351310|sp|P43496|TRXB_PENCH THIOREDOXIN REDUCTASE")
         self.assertEqual(record.rounds[0].alignments[13].length, 334)
-
-    def _check_text_2010L_phiblast_002_round1(self, record):
         self.assertEqual(len(record.rounds[1].new_seqs), 18)
         self.assertEqual(record.rounds[1].new_seqs[0].title, "gi|113709|sp|P27017|AMIC_PSEAE ALIPHATIC AMIDASE EXPRESSION-REG...")
         self.assertEqual(record.rounds[1].new_seqs[0].score, 49)
@@ -915,8 +904,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[22].title, ">gi|1175674|sp|P45116|YCIH_HAEIN HYPOTHETICAL PROTEIN HI1225")
         self.assertEqual(record.rounds[1].alignments[22].length, 106)
         self.assertEqual(record.rounds[1].alignments[23].title, ">gi|3025270|sp|P77269|YPHF_ECOLI ABC TRANSPORTER PERIPLASMIC BINDING PROTEIN YPHF PRECURSOR")
-
-    def _check_text_2010L_phiblast_002_round2(self, record):
         self.assertEqual(record.rounds[1].alignments[23].length, 327)
         self.assertEqual(len(record.rounds[2].new_seqs), 16)
         self.assertEqual(record.rounds[2].new_seqs[0].title, "gi|3024134|sp|O15303|MGR6_HUMAN METABOTROPIC GLUTAMATE RECEPTOR...")
@@ -1014,8 +1001,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[2].alignments[21].length, 725)
         self.assertEqual(record.rounds[2].alignments[22].title, ">gi|127971|sp||NCPR_SALTR_1 [Segment 1 of 3] NADPH-CYTOCHROME P450 REDUCTASE (CPR)")
         self.assertEqual(record.rounds[2].alignments[22].length, 426)
-
-    def _check_text_2010L_phiblast_002_hsps(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].score, 1897)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].bits, 743)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].expect, 0.0)
@@ -1432,8 +1417,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].identities, (28, 146))
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].positives, (53, 146))
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].gaps, (8, 146))
-
-    def _check_text_2010L_phiblast_002_hsps_details(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].query, "MKRKAKTIIAGIVALAVSQGAMADDIKVAIVGAMSGPVAQWGDMEFNGARQAIKDINAKGGIKGDKLVGVEYDDACDPKQAVAVANKIVNDGIQYVIGHLCSSSTQPASDIYEDEGILMISPGATNPELTQRGYQYIMRTAGLDSSQGPTAAKYILETVKPQRIAIIHDKQQYGEGLARSVQDGLKQGNANIVFFDGITAGEKDFSALIARLQKENIDFVYYGGYYPEMGQIVRQARANGLKTQFMGPEGVGNASLSNIAGGAAEGMLVTMPKRYDQDPANKAIVEALKADKKDPSGPYVWITYAAVQSLATAMTRSASHRPLDLVKDLKANGADTVIGPLKWDEKGDLKGFEFGVFQWHADGSSTVAK")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].match, "MKRKAKTIIAGIVALAVSQGAMADDIKVAIVGAMSGPVAQWGDMEFNGARQAIKDINAKGGIKGDKLVGVEYDDACDPKQAVAVANKIVNDGIQYVIGHLCSSSTQPASDIYEDEGILMISPGATNPELTQRGYQYIMRTAGLDSSQGPTAAKYILETVKPQRIAIIHDKQQYGEGLARSVQDGLKQGNANIVFFDGITAGEKDFSALIARLQKENIDFVYYGGYYPEMGQIVRQARANGLKTQFMGPEGVGNASLSNIAGGAAEGMLVTMPKRYDQDPANKAIVEALKADKKDPSGPYVWITYAAVQSLATAMTRSASHRPLDLVKDLKANGADTVIGPLKWDEKGDLKGFEFGVFQWHADGSSTVAK")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].sbjct, "MKRKAKTIIAGIVALAVSQGAMADDIKVAIVGAMSGPVAQWGDMEFNGARQAIKDINAKGGIKGDKLVGVEYDDACDPKQAVAVANKIVNDGIQYVIGHLCSSSTQPASDIYEDEGILMISPGATNPELTQRGYQYIMRTAGLDSSQGPTAAKYILETVKPQRIAIIHDKQQYGEGLARSVQDGLKQGNANIVFFDGITAGEKDFSALIARLQKENIDFVYYGGYYPEMGQIVRQARANGLKTQFMGPEGVGNASLSNIAGGAAEGMLVTMPKRYDQDPANKAIVEALKADKKDPSGPYVWITYAAVQSLATAMTRSASHRPLDLVKDLKANGADTVIGPLKWDEKGDLKGFEFGVFQWHADGSSTVAK")
@@ -1861,8 +1844,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].query_end, 349)
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].sbjct_start, 15)
         self.assertEqual(record.rounds[2].alignments[22].hsps[0].sbjct_end, 157)
-
-    def _check_text_2010L_phiblast_002_footer(self, record):
         self.assertEqual(record.database_name, ["data/swissprot"])
         self.assertEqual(record.num_letters_in_database, [29652561])
         self.assertEqual(record.num_sequences_in_database, [82258])
@@ -1926,14 +1907,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.database_letters, 29652561)
         self.assertEqual(len(record.rounds), 2)
         self.assertEqual(len(record.rounds[0].new_seqs), 30)
-        # Rest of test broken up to avoid Jython JVM limitations
-        self._check_text_2010L_phiblast_003_round0(record)
-        self._check_text_2010L_phiblast_003_round1(record)
-        self._check_text_2010L_phiblast_003_hsps(record)
-        self._check_text_2010L_phiblast_003_hsps_details(record)
-        self._check_text_2010L_phiblast_003_footer(record)
-
-    def _check_text_2010L_phiblast_003_round0(self, record):
         self.assertEqual(record.rounds[0].new_seqs[0].title, "gi|399896|sp|Q02134|HIS7_LACLA IMIDAZOLEGLYCEROL-PHOSPHATE DEHY...")
         self.assertEqual(record.rounds[0].new_seqs[0].score, 409)
         self.assertAlmostEqual(record.rounds[0].new_seqs[0].e, 1.e-114)
@@ -2085,8 +2058,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[0].alignments[28].length, 956)
         self.assertEqual(record.rounds[0].alignments[29].title, ">gi|1174406|sp|P36126|SP14_YEAST PHOSPHOLIPASE D1 (PLD 1) (CHOLINE PHOSPHATASE 1) (PHOSPHATIDYLCHOLINE-HYDROLYZING PHOSPHOLIPASE D1) (MEIOSIS-SPECIFIC SPORULATION PROTEIN SPO14)")
         self.assertEqual(record.rounds[0].alignments[29].length, 1380)
-
-    def _check_text_2010L_phiblast_003_round1(self, record):
         self.assertEqual(len(record.rounds[1].new_seqs), 0)
         self.assertEqual(len(record.rounds[1].alignments), 24)
         self.assertEqual(record.rounds[1].alignments[0].title, ">gi|2495230|sp|Q43072|HIS7_PEA IMIDAZOLEGLYCEROL-PHOSPHATE DEHYDRATASE (IGPD)")
@@ -2138,8 +2109,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[23].title, ">gi|3334215|sp|O33773|HIS7_SULSO PROBABLE IMIDAZOLEGLYCEROL-PHOSPHATE DEHYDRATASE (IGPD)")
         self.assertEqual(record.rounds[1].alignments[23].length, 193)
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].score, 1040)
-
-    def _check_text_2010L_phiblast_003_hsps(self, record):
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].bits, 409)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].expect, 1e-114)
         self.assertEqual(len(record.rounds[0].alignments[0].hsps), 1)
@@ -2514,8 +2483,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].identities, (83, 200))
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].positives, (124, 200))
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].gaps, (7, 200))
-
-    def _check_text_2010L_phiblast_003_hsps_details(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].query, "MTRISHITRNTKETQIELSINLDGTGQADISTGIGFLDHMLTLLTFHSDFDLKIIGHGDHETVGMDPHHLIEDVAIALGKCISEDLGNKLGIRRYGSFTIPMDEALVTCDLDISGRPYLVFHADLSGNQKLGGYDTEMTEEFFRALAFNAGITLHLNEHYGQNTHHIIEGMFKSTARALKQAVSIDESKVGEIPSSKGVL")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].match, "MTRISHITRNTKETQIELSINLDGTGQADISTGIGFLDHMLTLLTFHSDFDLKIIGHGDHETVGMDPHHLIEDVAIALGKCISEDLGNKLGIRRYGSFTIPMDEALVTCDLDISGRPYLVFHADLSGNQKLGGYDTEMTEEFFRALAFNAGITLHLNEHYGQNTHHIIEGMFKSTARALKQAVSIDESKVGEIPSSKGVL")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].sbjct, "MTRISHITRNTKETQIELSINLDGTGQADISTGIGFLDHMLTLLTFHSDFDLKIIGHGDHETVGMDPHHLIEDVAIALGKCISEDLGNKLGIRRYGSFTIPMDEALVTCDLDISGRPYLVFHADLSGNQKLGGYDTEMTEEFFRALAFNAGITLHLNEHYGQNTHHIIEGMFKSTARALKQAVSIDESKVGEIPSSKGVL")
@@ -2894,8 +2861,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].query_end, 200)
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].sbjct_start, 1)
         self.assertEqual(record.rounds[1].alignments[23].hsps[0].sbjct_end, 193)
-
-    def _check_text_2010L_phiblast_003_footer(self, record):
         self.assertEqual(record.database_name, ["data/swissprot"])
         self.assertEqual(record.num_letters_in_database, [29652561])
         self.assertEqual(record.num_sequences_in_database, [82258])
@@ -5668,14 +5633,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.database_letters, 29652561)
         self.assertEqual(len(record.rounds), 2)
         self.assertEqual(len(record.rounds[0].new_seqs), 30)
-        # Rest of test broken up to avoid Jython JVM limitations
-        self._check_text_2011L_psiblast_002_round0(record)
-        self._check_text_2011L_psiblast_002_round1(record)
-        self._check_text_2011L_psiblast_002_hsps(record)
-        self._check_text_2011L_psiblast_002_hsps_details(record)
-        self._check_text_2011L_psiblast_002_footer(record)
-
-    def _check_text_2011L_psiblast_002_round0(self, record):
         self.assertEqual(record.rounds[0].new_seqs[0].title, "gi|399896|sp|Q02134|HIS7_LACLA IMIDAZOLEGLYCEROL-PHOSPHATE DEHY...")
         self.assertEqual(record.rounds[0].new_seqs[0].score, 409)
         self.assertAlmostEqual(record.rounds[0].new_seqs[0].e, 1e-114)
@@ -5827,8 +5784,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[0].alignments[28].length, 1380)
         self.assertEqual(record.rounds[0].alignments[29].title, ">gi|3287848|sp|Q16099|GLK4_HUMAN GLUTAMATE RECEPTOR, IONOTROPIC KAINATE 4 PRECURSOR (GLUTAMATE RECEPTOR KA-1) (KA1) (EXCITATORY AMINO ACID RECEPTOR 1) (EAA1)")
         self.assertEqual(record.rounds[0].alignments[29].length, 956)
-
-    def _check_text_2011L_psiblast_002_round1(self, record):
         self.assertEqual(len(record.rounds[1].new_seqs), 2)
         self.assertEqual(record.rounds[1].new_seqs[0].title, "gi|2833252|sp|Q14571|IP3S_HUMAN INOSITOL 1,4,5-TRISPHOSPHATE-BI...")
         self.assertEqual(record.rounds[1].new_seqs[0].score, 30)
@@ -5889,8 +5844,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[24].length, 2701)
         self.assertEqual(record.rounds[1].alignments[25].title, ">gi|266389|sp|P29995|IP3S_RAT INOSITOL 1,4,5-TRISPHOSPHATE-BINDING PROTEIN TYPE 2 RECEPTOR (TYPE 2 INSP3 RECEPTOR) (TYPE 2 INOSITOL 1,4,5-TRISPHOSPHATE RECEPTOR)")
         self.assertEqual(record.rounds[1].alignments[25].length, 2701)
-
-    def _check_text_2011L_psiblast_002_hsps(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].score, 1040)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].bits, 409)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].expect, 1e-114)
@@ -6114,8 +6067,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertAlmostEqual(record.rounds[1].alignments[25].hsps[0].bits, 28.8)
         self.assertAlmostEqual(record.rounds[1].alignments[25].hsps[0].expect, 8.2)
         self.assertEqual(len(record.rounds[1].alignments[25].hsps), 1)
-
-    def _check_text_2011L_psiblast_002_hsps_details(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].identities, (200, 200))
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].positives, (200, 200))
         self.assertEqual(record.rounds[0].alignments[1].hsps[0].identities, (99, 198))
@@ -6674,8 +6625,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[25].hsps[0].query_end, 141)
         self.assertEqual(record.rounds[1].alignments[25].hsps[0].sbjct_start, 1540)
         self.assertEqual(record.rounds[1].alignments[25].hsps[0].sbjct_end, 1659)
-
-    def _check_text_2011L_psiblast_002_footer(self, record):
         self.assertEqual(record.database_name, ["data/swissprot"])
         self.assertEqual(record.num_letters_in_database, [29652561])
         self.assertEqual(record.num_sequences_in_database, [82258])
@@ -9994,17 +9943,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.database_sequences, 88201)
         self.assertEqual(record.database_letters, 31957340)
         self.assertEqual(len(record.rounds), 5)
-        # Rest of test broken up to avoid Jython JVM limitations
-        self._check_text_2012L_psiblast_001_round0(record)
-        self._check_text_2012L_psiblast_001_round1(record)
-        self._check_text_2012L_psiblast_001_round2(record)
-        self._check_text_2012L_psiblast_001_round4(record)
-        self._check_text_2012L_psiblast_001_hsps(record)
-        self._check_text_2012L_psiblast_001_hsps_counts(record)
-        self._check_text_2012L_psiblast_001_hsps_details(record)
-        self._check_text_2012L_psiblast_001_footer(record)
-
-    def _check_text_2012L_psiblast_001_round0(self, record):
         self.assertEqual(len(record.rounds[0].new_seqs), 27)
         self.assertEqual(record.rounds[0].new_seqs[0].title, "100K_RAT Q62671 rattus norvegicus (rat). 100 kda protein (ec...")
         self.assertEqual(record.rounds[0].new_seqs[0].score, 1516)
@@ -10142,8 +10080,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[0].alignments[25].length, 211)
         self.assertEqual(record.rounds[0].alignments[26].title, ">CC24_YEAST P11433 saccharomyces cerevisiae (baker's yeast). cell division control protein 24 (calcium regulatory protein). 7/1999")
         self.assertEqual(record.rounds[0].alignments[26].length, 854)
-
-    def _check_text_2012L_psiblast_001_round1(self, record):
         self.assertEqual(len(record.rounds[1].new_seqs), 9)
         self.assertEqual(record.rounds[1].new_seqs[0].title, "PABP_DROME P21187 drosophila melanogaster (fruit fly). polya...")
         self.assertEqual(record.rounds[1].new_seqs[0].score, 67)
@@ -10225,8 +10161,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[1].alignments[24].length, 863)
         self.assertEqual(record.rounds[1].alignments[25].title, ">SYQ_YEAST P13188 saccharomyces cerevisiae (baker's yeast). glutaminyl-trna synthetase (ec 6.1.1.18) (glutamine--trna ligase) (glnrs). 11/1997")
         self.assertEqual(record.rounds[1].alignments[25].length, 809)
-
-    def _check_text_2012L_psiblast_001_round2(self, record):
         self.assertEqual(len(record.rounds[2].new_seqs), 6)
         self.assertEqual(record.rounds[2].new_seqs[0].title, "PAB2_ARATH P42731 arabidopsis thaliana (mouse-ear cress). po...")
         self.assertEqual(record.rounds[2].new_seqs[0].score, 48)
@@ -10295,8 +10229,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[2].alignments[22].length, 892)
         self.assertEqual(record.rounds[2].alignments[23].title, ">RNR_AQUAE O67834 aquifex aeolicus. ribonuclease r (ec 3.1.-.-) (rnase r) (vacb protein homolog). 5/2000")
         self.assertEqual(record.rounds[2].alignments[23].length, 705)
-
-    def _check_text_2012L_psiblast_001_round3(self, record):
         self.assertEqual(len(record.rounds[3].new_seqs), 4)
         self.assertEqual(record.rounds[3].new_seqs[0].title, "PAB5_ARATH Q05196 arabidopsis thaliana (mouse-ear cress). po...")
         self.assertEqual(record.rounds[3].new_seqs[0].score, 51)
@@ -10359,8 +10291,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[3].alignments[22].length, 705)
         self.assertEqual(record.rounds[3].alignments[23].title, ">NIRA_EMENI P28348 emericella nidulans (aspergillus nidulans). nitrogen assimilation transcription factor nira. 4/1993")
         self.assertEqual(record.rounds[3].alignments[23].length, 892)
-
-    def _check_text_2012L_psiblast_001_round4(self, record):
         self.assertEqual(len(record.rounds[4].new_seqs), 2)
         self.assertEqual(record.rounds[4].new_seqs[0].title, "RNR_AQUAE O67834 aquifex aeolicus. ribonuclease r (ec 3.1.-....")
         self.assertEqual(record.rounds[4].new_seqs[0].score, 33)
@@ -10417,8 +10347,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[4].alignments[22].length, 705)
         self.assertEqual(record.rounds[4].alignments[23].title, ">NIRA_EMENI P28348 emericella nidulans (aspergillus nidulans). nitrogen assimilation transcription factor nira. 4/1993")
         self.assertEqual(record.rounds[4].alignments[23].length, 892)
-
-    def _check_text_2012L_psiblast_001_hsps(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].score, 3882)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].bits, 1516)
         self.assertAlmostEqual(record.rounds[0].alignments[0].hsps[0].expect, 0.0)
@@ -10930,8 +10858,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertAlmostEqual(record.rounds[4].alignments[23].hsps[0].bits, 32.9)
         self.assertAlmostEqual(record.rounds[4].alignments[23].hsps[0].expect, 2.9)
         self.assertEqual(len(record.rounds[4].alignments[23].hsps), 1)
-
-    def _check_text_2012L_psiblast_001_hsps_counts(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].identities, (765, 889))
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].positives, (765, 889))
         self.assertEqual(record.rounds[0].alignments[1].hsps[0].identities, (281, 634))
@@ -11281,8 +11207,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[4].alignments[22].hsps[0].gaps, (13, 92))
         self.assertEqual(record.rounds[4].alignments[23].hsps[0].identities, (22, 95))
         self.assertEqual(record.rounds[4].alignments[23].hsps[0].positives, (35, 95))
-
-    def _check_text_2012L_psiblast_001_hsps_details(self, record):
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].query, "MMSARGDFLNYALSLMRSHNDEHSDVLPVLDVCSLKHVAYVFQALIYWIKAMNQQTTLDTPQXXXXXXXXXXXXGIXXXXXXXXXXXXTSQSATLNDKDDESLPAETGQNHPFFRRSDSMTFLGCIPPNPFEVPLAEAIPLADQPHLLQPNARKEDLFGRPSQGLYSSSAGSGKCLVEVTMDRNCLEVLPTKMSYAANLKNVMNMQNRQKKAGEDQSMLAEEADSSKPGPSAHDVAAQLKSSLLAEIGLTESEGPPLTSFRPQCSFMGMVISHDMLLGRWRLSLELFGRVFMEDVGAEPGSILTELGGFEVKESKFRREMEKLRNQQSRDLSLEVDRDRDLLIQQTMRQLNNHFGRRCATTPMAVHRVKVTFKDEPGEGSGVARSFYTAIAQAFLSNEKLPNLDCIQNANKGTHTSLMQXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXQLSIDTRPFRPASEGNPSDDPDPLPAHRQALGERLYPRVQAMQPAFASKITGMXXXXXXXXXXXXXXXXXXXRARVEEAMELIVAHGRENGAXXXXXXXXXXXXEKVQENRKRHGSSRSVVXXXXXXXXXXXXNAPLFYQPGKRGFYTPRPGKNTEARLNCFRNIGRILGLCLLQNELCPITLNRHVIKVLLGRKVNWHDFAFFDPVMYESLRQLILASQSSDADAVFSAMDLAFAVDLCKEEGGGQVELIPNGVNIPVTPQNVYEYVRKYAEHRMLVVAEQPLHAMRKGLLDVLPKNSLEDLTAEDFRLLVNGCGEVNVQMLISFTSFNDESGENAEKLLQFKRWFWSIVERMSMTERQDLVYFWTSSPSLPASEEGFQPMPSITIRPPDDQHLPTANTCISRLYVPXXXXXXXXXXXXXXXXXXXNFGFV")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].match, "MMSARGDFLNYALSLMRSHNDEHSDVLPVLDVCSLKHVAYVFQALIYWIKAMNQQTTLDTPQ            GI            TSQSATLNDKDDESLPAETGQNHPFFRRSDSMTFLGCIPPNPFEVPLAEAIPLADQPHLLQPNARKEDLFGRPSQGLYSSSAGSGKCLVEVTMDRNCLEVLPTKMSYAANLKNVMNMQNRQKKAGEDQSMLAEEADSSKPGPSAHDVAAQLKSSLLAEIGLTESEGPPLTSFRPQCSFMGMVISHDMLLGRWRLSLELFGRVFMEDVGAEPGSILTELGGFEVKESKFRREMEKLRNQQSRDLSLEVDRDRDLLIQQTMRQLNNHFGRRCATTPMAVHRVKVTFKDEPGEGSGVARSFYTAIAQAFLSNEKLPNLDCIQNANKGTHTSLMQ                                      QLSIDTRPFRPASEGNPSDDPDPLPAHRQALGERLYPRVQAMQPAFASKITGM                   RARVEEAMELIVAHGRENGA            EKVQENRKRHGSSRSVV            NAPLFYQPGKRGFYTPRPGKNTEARLNCFRNIGRILGLCLLQNELCPITLNRHVIKVLLGRKVNWHDFAFFDPVMYESLRQLILASQSSDADAVFSAMDLAFAVDLCKEEGGGQVELIPNGVNIPVTPQNVYEYVRKYAEHRMLVVAEQPLHAMRKGLLDVLPKNSLEDLTAEDFRLLVNGCGEVNVQMLISFTSFNDESGENAEKLLQFKRWFWSIVERMSMTERQDLVYFWTSSPSLPASEEGFQPMPSITIRPPDDQHLPTANTCISRLYVP                   NFGFV")
         self.assertEqual(record.rounds[0].alignments[0].hsps[0].sbjct, "MMSARGDFLNYALSLMRSHNDEHSDVLPVLDVCSLKHVAYVFQALIYWIKAMNQQTTLDTPQLERKRTRELLELGIDNEDSEHENDDDTSQSATLNDKDDESLPAETGQNHPFFRRSDSMTFLGCIPPNPFEVPLAEAIPLADQPHLLQPNARKEDLFGRPSQGLYSSSAGSGKCLVEVTMDRNCLEVLPTKMSYAANLKNVMNMQNRQKKAGEDQSMLAEEADSSKPGPSAHDVAAQLKSSLLAEIGLTESEGPPLTSFRPQCSFMGMVISHDMLLGRWRLSLELFGRVFMEDVGAEPGSILTELGGFEVKESKFRREMEKLRNQQSRDLSLEVDRDRDLLIQQTMRQLNNHFGRRCATTPMAVHRVKVTFKDEPGEGSGVARSFYTAIAQAFLSNEKLPNLDCIQNANKGTHTSLMQRLRNRGERDREREREREMRRSSGLRAGSRRDRDRDFRRQLSIDTRPFRPASEGNPSDDPDPLPAHRQALGERLYPRVQAMQPAFASKITGMLLELSPAQLLLLLASEDSLRARVEEAMELIVAHGRENGADSILDLGLLDSSEKVQENRKRHGSSRSVVDMDLDDTDDGDDNAPLFYQPGKRGFYTPRPGKNTEARLNCFRNIGRILGLCLLQNELCPITLNRHVIKVLLGRKVNWHDFAFFDPVMYESLRQLILASQSSDADAVFSAMDLAFAVDLCKEEGGGQVELIPNGVNIPVTPQNVYEYVRKYAEHRMLVVAEQPLHAMRKGLLDVLPKNSLEDLTAEDFRLLVNGCGEVNVQMLISFTSFNDESGENAEKLLQFKRWFWSIVERMSMTERQDLVYFWTSSPSLPASEEGFQPMPSITIRPPDDQHLPTANTCISRLYVPLYSSKQILKQKLLLAIKTKNFGFV")
@@ -12193,8 +12117,6 @@ class TestNCBITextParser(unittest.TestCase):
         self.assertEqual(record.rounds[4].alignments[23].hsps[0].query_end, 313)
         self.assertEqual(record.rounds[4].alignments[23].hsps[0].sbjct_start, 702)
         self.assertEqual(record.rounds[4].alignments[23].hsps[0].sbjct_end, 796)
-
-    def _check_text_2012L_psiblast_001_footer(self, record):
         self.assertEqual(record.database_name, ["/dbase/swissprot/main/release/sp", "/dbase/swissprot/main/update/spu"])
         self.assertEqual(record.posted_date, [("Jun 21, 2000 12:39 PM",), ("Nov 3, 1999  8:09 PM",)])
         self.assertEqual(record.num_letters_in_database, [31411157, 546183])
